Fix: Recruit 페이지 코드리뷰 반영 #9
