+++ next.config.mjs
@@ -1,4 +1,13 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  webpack: config => {
+    config.module.rules.push({
+      test: /\.svg$/i,
+      use: ['@svgr/webpack']
+    })
 
-export default nextConfig;
+    return config
+  }
+}
+
+export default nextConfig
