+++ package.json
@@ -28,6 +28,7 @@
     "react-lottie-player": "^2.1.0"
   },
   "devDependencies": {
+    "@svgr/webpack": "^8.1.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
