+++ src/app/recruit/apply/container/SaveComplete.tsx
@@ -1,0 +1,35 @@
+import React from 'react'
+import LikelionGreyLogo from '../../../../../public/icon/logo/likelion-grey.svg'
+
+const SaveComplete = () => {
+  return (
+    <div className="mt-[24.7rem] flex flex-col items-center">
+      <div>
+        <LikelionGreyLogo />
+      </div>
+      <div className="mt-[3.4rem] text-[1.6rem] font-bold">
+        김경우 님, 지원서 저장이 완료되었습니다.
+      </div>
+      <div className="text-[1.2rem] font-medium">
+        likelion@sogang.ac.kr로 발송된 고유 번호를 입력하여 지원서를 수정할 수
+        있습니다.
+      </div>
+      <div className="mt-[1.2rem] text-[#b7b7b7] text-[1.2rem] font-semibold">
+        최종 제출 시각: {new Date().toLocaleDateString()}
+        <span> </span>
+        {new Date().toLocaleTimeString()}
+      </div>
+
+      <div className="flex mt-[5.8rem] gap-[0.8rem]">
+        <div className="p-[0.6rem_1.3rem] bg-grey text-[#b7b7b7] text-[1.3rem] font-semibold rounded-[2rem]">
+          홈으로 돌아가기
+        </div>
+        <div className="p-[0.6rem_1.3rem] bg-grey text-[#b7b7b7] text-[1.3rem] font-semibold rounded-[2rem]">
+          지원서 수정하기
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default SaveComplete
