+++ src/app/recruit/apply/page.tsx
@@ -1,0 +1,9 @@
+import ApplySection from './container/ApplySection'
+
+export default function ApplyPage() {
+  return (
+    <div>
+      <ApplySection />
+    </div>
+  )
+}
