--- src/app/recruit/container/Nonpassed.tsx
@@ -1,57 +1,0 @@
-import React from 'react'
-
-const Nonpassed = () => {
-  return (
-    <div className="text-black flex flex-col justify-center items-center w-[56.2rem]">
-      <div className="mb-[3.839rem]">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="130"
-          height="90"
-          viewBox="0 0 130 90"
-          fill="none">
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M65.004 85.6574C65.0114 86.9783 65.004 88.2952 64.9815 89.6084C62.0791 88.1497 59.1832 86.6793 56.2939 85.1973C53.0206 83.4923 49.8403 81.6441 46.7532 79.6524C44.4964 78.1744 42.3413 76.5581 40.2879 74.8035C38.4608 73.2159 36.7884 71.4836 35.2707 69.6066C31.9673 65.4128 30.2163 60.6312 30.0177 55.2619C29.883 50.9973 30.5189 46.8519 31.9258 42.8253C33.5468 38.7632 34.744 34.5801 35.5176 30.2765C36.9554 21.9725 36.7159 13.7188 34.7992 5.51556C34.3559 3.6503 33.8059 1.81713 33.1493 0.015625C34.4663 0.015625 35.7832 0.015625 37.1002 0.015625C37.7014 2.7638 38.2815 5.51735 38.84 8.27675C39.3072 10.6294 39.6365 13.0015 39.8277 15.393C40.1164 20.5708 39.7647 25.7039 38.7727 30.7928C38.3026 33.1373 37.7337 35.457 37.0665 37.7519C36.1132 40.8842 35.1703 44.0196 34.238 47.1579C34.0566 47.7916 33.8844 48.4278 33.7217 49.0661C33.66 49.3336 33.6151 49.603 33.587 49.8742C44.0629 39.7049 54.5316 29.5283 64.9927 19.344C65.0078 24.6194 65.0078 29.8949 64.9927 35.1703C55.1768 44.4353 45.3629 53.7028 35.5513 62.973C35.4752 63.0179 35.4415 63.0815 35.4502 63.1638C36.4209 65.2029 37.6482 67.0661 39.1318 68.7536C40.3351 70.129 41.6333 71.4048 43.0267 72.5811C45.4815 74.6167 48.0705 76.4649 50.7939 78.1259C53.7572 79.9335 56.8028 81.6097 59.9306 83.1544C60.9033 83.596 61.8762 84.0373 62.8489 84.4789C63.5933 84.8318 64.3116 85.2246 65.004 85.6574Z"
-            fill="black"
-          />
-          <path
-            d="M0 0H37.1147L38.1341 24.2424L0 0Z"
-            fill="black"
-          />
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M64.9969 85.6574C64.9895 86.9783 64.9969 88.2952 65.0194 89.6084C67.9218 88.1497 70.8176 86.6793 73.707 85.1973C76.9803 83.4923 80.1606 81.6441 83.2477 79.6524C85.5045 78.1744 87.6596 76.5581 89.7129 74.8035C91.54 73.2159 93.2125 71.4836 94.7302 69.6066C98.0336 65.4128 99.7846 60.6312 99.9832 55.2619C100.118 50.9973 99.4819 46.8519 98.0751 42.8253C96.4541 38.7632 95.2569 34.5801 94.4833 30.2765C93.0455 21.9725 93.285 13.7188 95.2017 5.51556C95.645 3.6503 96.195 1.81713 96.8516 0.015625C95.5346 0.015625 94.2177 0.015625 92.9007 0.015625C92.2995 2.7638 91.7194 5.51735 91.1609 8.27675C90.6937 10.6294 90.3644 13.0015 90.1731 15.393C89.8845 20.5708 90.2362 25.7039 91.2282 30.7928C91.6983 33.1373 92.2672 35.457 92.9343 37.7519C93.8877 40.8842 94.8306 44.0196 95.7629 47.1579C95.9443 47.7916 96.1164 48.4278 96.2792 49.0661C96.3409 49.3336 96.3858 49.603 96.4139 49.8742C85.938 39.7049 75.4692 29.5283 65.0081 19.344C64.9931 24.6194 64.9931 29.8949 65.0081 35.1703C74.8241 44.4353 84.638 53.7028 94.4496 62.973C94.5257 63.0179 94.5594 63.0815 94.5506 63.1638C93.58 65.2029 92.3527 67.0661 90.8691 68.7536C89.6658 70.129 88.3676 71.4048 86.9742 72.5811C84.5194 74.6167 81.9304 76.4649 79.2069 78.1259C76.2437 79.9335 73.1981 81.6097 70.0703 83.1544C69.0976 83.596 68.1247 84.0373 67.152 84.4789C66.4076 84.8318 65.6892 85.2246 64.9969 85.6574Z"
-            fill="#B2B9BC"
-          />
-          <path
-            d="M130 0H92.8854L91.8662 24.2424L110.682 12.2808L130 0Z"
-            fill="#B2B9BC"
-          />
-        </svg>
-      </div>
-
-      <div className="text-[2.4rem] font-pretendard font-semibold text-center mb-[6.5rem] whitespace-pre-line">
-        {'김경우 님, 멋쟁이사자처럼 서강대학교 12기\n 최종 합격을 축하드려요!'}
-      </div>
-
-      <div className="text-center text-[1.6rem] font-medium font-pretendard whitespace-pre-line leading-[23px]">
-        {
-          '안녕하세요, 멋쟁이사자처럼 서강대학교 12기 운영진입니다.\n12기 멋쟁이사자처럼 서강대학교 [최종 합격]을 축하드립니다.\n\n그럼 12기 아기사자 활동을 시작하기에 앞서 첫 번째 공지사항 전달드립니다.\n\n1. 정규 세션시간: 매주 화요일 / 금요일 19:00~21:00\n\n2. 오리엔테이션: 3월 19일 (화요일) 19:00 마포 프론트원 - 공덕 ICT COC\n\n3. 강의 자료, 과제 등의 업로드는 Notion과 Github를 사용합니다. Notion에 가입해주시고, Notion에 가입 이메일을 아래 구글폼에 적어서 제출해주시기 바랍니다.\n\n4. 멋쟁이사자처럼 서강대학교는 회칙 제 11조에 따라 본회의 유지 및 운영에 필요한 경비 5만원을 회비로 정하고 있습니다. \n\n12기 총무계좌는 아래와 같습니다.\n3333291762013 카카오뱅크 정고은\n\n5. 멋쟁이사자처럼 12기 회원 정보를 수합합니다. 추가로 회비 입금 내역, 이후 일정 참가에 대하여 아래 구글폼을 입력해주시면 감사하겠습니다. 구글폼은 3/19(일) 오전 11:59(정오)까지 제출 부탁드립니다. \n\n  '
-        }
-        <a
-          className="underline"
-          href="https://forms.gle/YrDqQ1rSX3UBfvPh8">
-          https://forms.gle/YrDqQ1rSX3UBfvPh8
-        </a>
-        {
-          '\n\n다시 한번 멋쟁이사자처럼 서강대학교의 일원이 되신 것을 축하드리며, 앞으로 여러분과 함께 하게 될 수많은 시간들을 기대하고 있겠습니다! \n\n🦁 POSSIBILITY TO REALITY 🦁'
-        }
-      </div>
-    </div>
-  )
-}
-
-export default Nonpassed
