--- src/app/recruit/container/SaveComplete.tsx
@@ -1,60 +1,0 @@
-import React from 'react'
-
-const SaveComplete = () => {
-  return (
-    <div className="mt-[24.7rem] flex flex-col items-center">
-      <div>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="130"
-          height="90"
-          viewBox="0 0 130 90"
-          fill="none">
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M65.0045 85.6574C65.0119 86.9783 65.0045 88.2952 64.982 89.6084C62.0796 88.1497 59.1837 86.6793 56.2944 85.1973C53.0211 83.4923 49.8408 81.6441 46.7537 79.6524C44.4969 78.1744 42.3418 76.5581 40.2884 74.8035C38.4613 73.2159 36.7889 71.4836 35.2711 69.6066C31.9678 65.4128 30.2168 60.6312 30.0181 55.2619C29.8835 50.9973 30.5194 46.8519 31.9263 42.8253C33.5473 38.7632 34.7445 34.5801 35.5181 30.2765C36.9559 21.9725 36.7164 13.7188 34.7997 5.51556C34.3564 3.6503 33.8064 1.81713 33.1497 0.015625C34.4668 0.015625 35.7836 0.015625 37.1007 0.015625C37.7019 2.7638 38.282 5.51735 38.8405 8.27675C39.3076 10.6294 39.637 13.0015 39.8282 15.393C40.1169 20.5708 39.7651 25.7039 38.7731 30.7928C38.3031 33.1373 37.7342 35.457 37.067 37.7519C36.1136 40.8842 35.1708 44.0196 34.2385 47.1579C34.0571 47.7916 33.8849 48.4278 33.7222 49.0661C33.6604 49.3336 33.6155 49.603 33.5875 49.8742C44.0634 39.7049 54.5321 29.5283 64.9932 19.344C65.0083 24.6194 65.0083 29.8949 64.9932 35.1703C55.1773 44.4353 45.3634 53.7028 35.5518 62.973C35.4756 63.0179 35.442 63.0815 35.4507 63.1638C36.4214 65.2029 37.6487 67.0661 39.1323 68.7536C40.3356 70.129 41.6338 71.4048 43.0272 72.5811C45.4819 74.6167 48.0709 76.4649 50.7944 78.1259C53.7577 79.9335 56.8033 81.6097 59.931 83.1544C60.9038 83.596 61.8767 84.0373 62.8494 84.4789C63.5938 84.8318 64.3121 85.2246 65.0045 85.6574Z"
-            fill="black"
-          />
-          <path
-            d="M0 0H37.1147L38.1341 24.2424L0 0Z"
-            fill="black"
-          />
-          <path
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M64.9964 85.6574C64.989 86.9783 64.9964 88.2952 65.0189 89.6084C67.9213 88.1497 70.8172 86.6793 73.7065 85.1973C76.9798 83.4923 80.1601 81.6441 83.2472 79.6524C85.504 78.1744 87.6591 76.5581 89.7125 74.8035C91.5396 73.2159 93.212 71.4836 94.7297 69.6066C98.0331 65.4128 99.7841 60.6312 99.9827 55.2619C100.117 50.9973 99.4815 46.8519 98.0746 42.8253C96.4536 38.7632 95.2564 34.5801 94.4828 30.2765C93.045 21.9725 93.2845 13.7188 95.2012 5.51556C95.6445 3.6503 96.1945 1.81713 96.8511 0.015625C95.5341 0.015625 94.2172 0.015625 92.9002 0.015625C92.299 2.7638 91.7189 5.51735 91.1604 8.27675C90.6932 10.6294 90.3639 13.0015 90.1727 15.393C89.884 20.5708 90.2357 25.7039 91.2277 30.7928C91.6978 33.1373 92.2667 35.457 92.9338 37.7519C93.8872 40.8842 94.8301 44.0196 95.7624 47.1579C95.9438 47.7916 96.116 48.4278 96.2787 49.0661C96.3404 49.3336 96.3853 49.603 96.4134 49.8742C85.9375 39.7049 75.4688 29.5283 65.0077 19.344C64.9926 24.6194 64.9926 29.8949 65.0077 35.1703C74.8236 44.4353 84.6375 53.7028 94.4491 62.973C94.5252 63.0179 94.5589 63.0815 94.5502 63.1638C93.5795 65.2029 92.3522 67.0661 90.8686 68.7536C89.6653 70.129 88.3671 71.4048 86.9737 72.5811C84.5189 74.6167 81.9299 76.4649 79.2065 78.1259C76.2432 79.9335 73.1976 81.6097 70.0698 83.1544C69.0971 83.596 68.1242 84.0373 67.1515 84.4789C66.4071 84.8318 65.6887 85.2246 64.9964 85.6574Z"
-            fill="#B2B9BC"
-          />
-          <path
-            d="M130 0H92.8854L91.8662 24.2424L110.682 12.2808L130 0Z"
-            fill="#B2B9BC"
-          />
-        </svg>
-      </div>
-      <div className="mt-[3.4rem] text-[1.6rem] font-bold">
-        김경우 님, 지원서 저장이 완료되었습니다.
-      </div>
-      <div className="text-[1.2rem] font-medium">
-        likelion@sogang.ac.kr로 발송된 고유 번호를 입력하여 지원서를 수정할 수
-        있습니다.
-      </div>
-      <div className="mt-[1.2rem] text-[#b7b7b7] text-[1.2rem] font-semibold">
-        최종 제출 시각: {new Date().toLocaleDateString()}
-        <span> </span>
-        {new Date().toLocaleTimeString()}
-      </div>
-
-      <div className="flex mt-[5.8rem] gap-[0.8rem]">
-        <div className="p-[0.6rem_1.3rem] bg-grey text-[#b7b7b7] text-[1.3rem] font-semibold rounded-[2rem]">
-          홈으로 돌아가기
-        </div>
-        <div className="p-[0.6rem_1.3rem] bg-grey text-[#b7b7b7] text-[1.3rem] font-semibold rounded-[2rem]">
-          지원서 수정하기
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default SaveComplete
