+++ src/app/recruit/page.tsx
@@ -2,21 +2,17 @@ import React from 'react'
 import Image from 'next/image'
 import Lottie from 'react-lottie-player'
 import animation from '../../../public/lottie/animation.json'
-import FirstSection from './container/FirstSection'
-import SecondSection from './container/SecondSection'
-import ApplySection from './container/ApplySection'
-import PersonalInformationForm from './container/PersonalInformationForm'
-import PersonalStatementForm from './container/PersonalStatementForm'
-import EditPersonalInfo from './container/EditPersonalInfo'
-import CheckingPass from './CheckingPass'
+import ApplySection from './apply/container/ApplySection'
+import EditPersonalInfo from './apply/container/EditPersonalInfo'
+import CheckingPass from './apply/CheckingPass'
+import Recruit1 from './container/Recruit1'
+import Recruit2 from './container/Recruit2'
 
 export default function RecruitPage() {
   return (
     <div>
-      <FirstSection />
-      <SecondSection />
-      <ApplySection />
-      <CheckingPass />
+      <Recruit1 />
+      <Recruit2 />
     </div>
   )
 }
