+++ src/custom.d.ts
@@ -1,0 +1,9 @@
+import { FC, SVGProps, SVGSVGElement } from 'react'
+
+declare global {
+  module '*.svg' {
+    const component: FC<SVGProps<SVGSVGElement>>
+
+    export default component
+  }
+}
